requirements.txt
SpeechRecognition~=3.10.4
openai
moviepy
openai-whisper
httpx[socks]